refactor(library): separate local implementation
